update 7 files and delete 1 file
