+++ src/rescuetime2gcal/sources/source_event.py
@@ -1,3 +1,4 @@
+"""A source event, i.e. right after ingest."""
 import datetime
 from abc import ABC
 from typing import TYPE_CHECKING
@@ -5,10 +6,9 @@ from typing import TYPE_CHECKING
 import pydantic
 
 if TYPE_CHECKING:
-    from rescuetime2gcal.event import ChronofileEvent, DestinationEvent
 
 
-class BaseEvent(pydantic.BaseModel, ABC):
+class BaseSourceEvent(pydantic.BaseModel, ABC):
     start: "datetime.datetime"
     duration: "datetime.timedelta"
 
@@ -20,7 +20,7 @@ class BaseEvent(pydantic.BaseModel, ABC):
         return f"[{self.start.isoformat(timespec='seconds')}, {self.duration.total_seconds() / 60:.2f}m] {description}"
 
 
-class BareEvent(BaseEvent):
+class BareEvent(BaseSourceEvent):
     title: str
 
     @pydantic.field_validator("title")
@@ -33,7 +33,7 @@ class BareEvent(BaseEvent):
         return super().repr_str(f"Bare, {self.title}")
 
 
-class URLEvent(BaseEvent):
+class URLEvent(BaseSourceEvent):
     url: str
     url_title: str
 
@@ -41,7 +41,7 @@ class URLEvent(BaseEvent):
         return super().repr_str(f"URL, {self.url_title}")
 
 
-class WindowTitleEvent(BaseEvent):
+class WindowTitleEvent(BaseSourceEvent):
     app: str
     window_title: str
 
@@ -49,10 +49,6 @@ class WindowTitleEvent(BaseEvent):
         return super().repr_str(f"Window, {self.app}: {self.window_title}")
 
 
-SourceEvent = BaseEvent | URLEvent | WindowTitleEvent | BareEvent
-"""A source event, i.e. right after ingest."""
+SourceEvent = BaseSourceEvent | URLEvent | WindowTitleEvent | BareEvent
 
 
-def event_identity(event: "DestinationEvent | ChronofileEvent") -> str:
-    string_format = "%d/%m/%Y, %H:%M:%S"
-    return f"{event.title} {event.start.strftime(string_format)} to {event.end.strftime(string_format)}"
